+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Version counting is based on semantic versioning (Major.Feature.Patch)
 * Add support for showing a "recently added/updated" indicator.
 * Improved comic metadata dialog.
 * Add textual tags support that can be queried through the search engine.
+* Make = in the search engine work as : does.
+* Add new operators to the search engine: exact match ==, <, >, <=, >=.
 
 ## 9.12